MessageHandler/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
amqpstorm==2.10.3
python-dotenv==1.0.1
python-multipart==0.0.9
twilio==8.12.0 
